+++ src/Shared/Extensions/ApplicationInsightsServiceOptions.cs
@@ -90,6 +90,11 @@
         public string InstrumentationKey { get; set; }
 
         /// <summary>
+        /// Gets or sets the connection string for the application.
+        /// </summary>
+        public string ConnectionString { get; set; }
+
+        /// <summary>
         /// Gets or sets the application version reported with telemetries.
         /// </summary>
         public string ApplicationVersion { get; set; }
@@ -164,6 +169,7 @@
                 target.InstrumentationKey = this.InstrumentationKey;
             }
 
+            target.ConnectionString = this.ConnectionString;
             target.ApplicationVersion = this.ApplicationVersion;
             target.EnableAdaptiveSampling = this.EnableAdaptiveSampling;
             target.EnableDebugLogger = this.EnableDebugLogger;